refactor: reorganize bot into package
