+++ bot/keyboards.py
@@ -3,7 +3,7 @@
 # ويولّد القوائم الديناميكية للمستويات، الأترام، المواد، الأقسام، الفلاتر، السنوات، المحاضرين، وعناوين المحاضرات.
 
 from telegram import ReplyKeyboardMarkup
-from config import ADMIN_USER_IDS
+from .config import ADMIN_USER_IDS
 # -----------------------------------------------------------------------------
 # ثوابت نصوص الأزرار (يجب أن تبقى متطابقة مع ما يستخدمه bot.py)
 # -----------------------------------------------------------------------------
